+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 Notable changes:
 
 * `bashy-core`:
+  * New helper function `set-array-from-lines`.
   * `lib`: New option `--exec`.
   * `ubik`: Finds functions from the library, not just commands (scripts).
 * `bashy-basics`: